pyproject.toml
[project]
name = "adam"
version = "0.1.0"
description = "LLM-powered robot motor cortex — speak natural language, get bone rotation keyframes"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.6",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "litellm>=1.83.8",
]

[project.scripts]
adam = "adam.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/adam"]
